Add dtype to Self

+++ code/modules.py
@@ -265,14 +265,14 @@ class SelfAttn(object):
         with vs.variable_scope("SelfAttn"):
 
             xi = tf.contrib.layers.xavier_initializer()
-            w1 = tf.get_variable("W1", shape=(self.key_vec_size, self.bahdanau_size), initializer=xi)
-            v = tf.get_variable("v", shape=(self.bahdanau_size), initializer=xi)
+            w1 = tf.get_variable("W1", shape=(self.key_vec_size, self.bahdanau_size), dtype = tf.float32, initializer=xi)
+            v = tf.get_variable("v", shape=(self.bahdanau_size), dtype = tf.float32, initializer=xi)
 
             # (batch_size, num_keys, bahdanau_size)
             keys_shape = keys.get_shape()
             w1_keys = tf.tensordot(keys, w1, axes=1)
             w1_keys.set_shape((keys_shape[0], keys_shape[1], self.bahdanau_size))
-            
+
             # (batch_size, num_keys, 1, bahdanau_size)
             w1_keys_ax2 = tf.expand_dims(w1_keys, axis=2)
